# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Uzima Platform API
  description: |
    Unified JWT + API-key backend for Uzima trade receivables (UZIMA-ARCH-001).
    Dual origination, AfyaX parties/webhooks, wallets, fees, documents, packages, programmes.
  version: 2.0.0
servers:
  - url: http://localhost:8787
    description: Local development
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
paths:
...
  /api/v1/auth/login:
    post:
      summary: Portal login (JWT + httpOnly refresh cookie)
      security: []
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: accessToken + user (refreshToken in cookie; body only if ALLOW_BODY_REFRESH)
        "401":
          description: Invalid credentials
        "429":
          description: Rate limited / locked out
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key